requirements.txt

aniso8601==8.0.0
APScheduler==3.6.3
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.1
cryptography==3.2.1
cymem==2.0.3
decorator==4.0.11
dnspython==1.16.0
feedparser==5.2.1
filetype==1.0.5
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
Flask-SocketIO==4.3.1
greenlet==0.4.16
idna==2.9
imageio==2.1.2
instagram-private-api==1.6.0.0
instagram-private-api-extensions==0.3.8
instapy-cli==0.0.12
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
MarkupSafe==1.1.1
mathparse==0.1.2
monotonic==1.5
moviepy==0.2.3.2
murmurhash==1.0.2
nltk==3.5
numpy==1.19.0
Pillow==8.1.0
plac==1.1.3
preshed==3.0.2
pycparser==2.20
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.0
python-telegram-bot==13.0
pytz==2019.3
PyYAML==5.3.1
regex==2020.10.28
requests==2.23.0
schedule==0.6.0
six==1.14.0
spacy==2.3.2
SQLAlchemy==1.3.20
srsly==1.0.2
thinc==7.4.1
tinydb==3.15.2
tornado==6.1
tqdm==4.11.2
tzlocal==2.1
urllib3==1.25.9
wasabi==0.7.0
Werkzeug==1.0.1
xmljson==0.2.1